Add working tables to table prefix conversion
The table prefixes within the two working SQL scripts can be adjusted
at the same time, and in the same process as they are adjusted in the
installer SQL.

Avoids the opportunity for the user to forget, or to do it incorrectly,
since the prefix is not used in all the table names. In the file that
is to be used within the Zen-Cart admin tools, `load_ag_files.sql` the
table names on the "TRUNCATE TABLE" lines does not have the prefix.
The query factory adds those before the patch is handed off to the
database. The table names which are the target of the "INFILE LOCAL"
command, however, are ignored by the query factory, and have to have
the prefix, if used, included.

As proof of how easy it would be for the user to not change those
files, just notice that the documentation also does not cover the need
to change the prefixes there.

Fixed a typo in INSTALL.md

Reversed the order of versions in the CHANGELOG files. I forgot that
the convention is for the newest to be at the top.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>