Change status detection
The data tables form AzureGreen have a column 'DISCONT' which was used
as "discontinued", meaning that the product was one they were going to
discontinue. The processing delay from orders here to orders shipping
is sufficient to justify having such items unavailable in the store.

The 'DISCONT' column is, instead, an indication that the item has a
"discount" on the price. There are no indications of that the amount
of the discount is, nor when it ends.

The change in meaning for that column requires a change in how the
importing logic determines the status of products, removing that from
the consideration.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>