Prepare for version 2
Discoveries in the level of corrections needed in processing the data
from AzureGreen, and the level of security imposed by some shared
hosting platforms, has created the need for a completely different,
and more thorough, processing of the data.

It was found that the data files and image names were significantly
more "dirty" than originally thought. This requires a reworking of
how the data is processed and how the image files are sorted.

It was also discovered that the available options on some shared
hosting platforms are significantly tighter than required for this
system to work. On key, among many, is the inability to do
    LOAD DATA LOCAL INFILE
commands in MySQL/MariaDB. This system will have to rely on
command-line access to the  database from a local machine. This can
be the database itself, or a local clone of the database.

Here begins the reworking for version 2.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>