Add routine to keep the larger of two images
Only will keep the larger of two images if there _are_ two images.
If there are two images, it will also remove the smaller one. Will
return true if there were two images and it kept one and removed the
other. Return of false means there was nothing to compare with.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>