Save cleaned data and import into database
Change the name of cleaned CSV files to avoid conflict with version 1.

If there was _any_ of the product data files found to be new, _all_
of the files need to be processed as a unit. The new files should be
built around the information in the other files even when they are
not updated.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>