Add the SQL script file
The file has no commands, yet. It is the plan for the script and is
all comments for now. It will be fleshed out as progress is made.

Again, like the Bash script, the vi folding markers are included, and
will be dropped in the final version. In this case, however, they
will not affect the execution of the script by MySQL.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>