Mark categories dropped by AzureGreen as inactive
The changes to the database need to be staged rather than going
directly into the active tables. The changes, if saved as they are
encountered, could affect processing later. In addition, the system
has to be able to update the database currently active in, and build
a patch script to update a possible remote database.

Tables for each set of changes will be created by the processing, and
once all processing is completed the tables will be used to both
update the current database and generate the patch script for remote
usage.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>