Change used columns in categories tables
The meta tag related fields generated by the process are just copied
from the one piece of data available, the name of the category in
AzureGreen's data. There's no reason to carry that through all the
processing. Nor is it needed from the live data. When categories are
created, or modified, the "generation" of the data can be handled.
The one field pair that matters is the category name (`dept_name`) of
AzureGreen, and `categories_description` in the Zen-Cart database.

Once modified locally, none of the fields are touched by the process,
and until then they are derived from `dept_name`. For processing, all
other fields are extra baggage.

Added some `KEY`s that seemed helpful to the tables and changed all
the temporary tables to use the MEMORY engine where possible. Changed
the `categories_description` columns from `TEXT` to `VARCHAR` as the
MEMORY engine cannot be used on tables with `TEXT` columns. Table
sizes might go beyond the MySQL limits for MEMORY engine usage, and be
written to disk in the background, but it may help performance
somewhat anyway.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>