Record problems and errors found in the data
Includes cases where, for new categories, the name supplied is too
long to fit in the Zen-Cart table. 32 characters is the Zen-Cart limit
while the AzureGreen data has a much higher limit, somewhere.

Cases where the parent category does not exist are a problem for the
system, since they don't handle well. These are reported here.

Also included are cases where a category is active, while it's parent
is inactive. This one is a lot of checking, since both new categories
and existing categories have to be compared as both the child and the
parent categories.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>