Convert the descriptions file last
Involves adding what isn't already there from the first file. The
stockinfo file has no descriptions, so it's not involved. Then update
for what was added from the ag_complete_files, but might be different
in this file.

The ag_complete_files file seems to be the least updated by AzureGreen
and is considered to be less correct than the other two.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>