Locate products with changed description or name
Using two separate tables, one for names and one for descriptions, as
it seems that AzureGreen commonly adjusts the description for products
while leaving the product name alone. There is no point to recording a
change of name, which didn't "change" every time there is an update to
the product description.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>