Add products_title to most of the product tables
As noted earlier, the `metatags_products_title` field is made from the
same data as the `products_name` field, which has been included in the
converted data. The two fields, however, are of different lengths and
the possibility exists that in chopping the source data for the name
field, some of the names could be shortened, while the full original
name would fit in the metatags field.

The option to simply extend the size of the name field, until it is
inserted seemed less than ideal as it makes the finding of chopped
names for the anomolies list a bit trickier. Adding a field for the
full-length title, where needed, seemed to be a better choice.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>