Report problems found with the new product data
So far the problems are products where the name from AzureGreen is too
long for the database, and products where either the description or
the name, or both, are missing.

Later processing will be set to find active products which have only
inactive categories to be listed under, and products which do not have
a category to be listed under.

Fixed a few typos discovered here and there. Including using TEXT
columns in tables with the MEMORY engine, which is not supported.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>