Update database with imported categories data
Updating the database involves updates to both the remote, or live,
system and the local tables, which are supposed to be copies of the
live tables, made for this process. The local copy of the tables are
able to be updated directly by the running script. The ability to
access the live database, presumed to be remote, is not assured.

At each stage of the table updates the elements of a script to be run
against the remote, "live" database are created. At the same point,
the changes are applied to the local copy of the database. In most
cases, the remote script elements are created prior to updating the
local copy. This allows for record selection based on comparisons
with the "current" data. Changing the local copy first would void
the same comparisons when creating the remote script.

There are situations, primarily when using AUTO_INCREMENT values, in
which the local copy needs to be updated first, and the remote script
created using literal values taken from the local tables. This ensures
that generated values are consistent across both copies of the tables.

Because of unknown limitations on various implementations of MySQL/
MariaDB, the script lines are generated as direct output from the
importing script, which is able to be saved as a file on the local
system, using redirection of `standard output`.  Even if the MySQL
command line client is accessing some remote host, the output will
still be available on the local system, where the user is presumed to
have full access to their working directory.

Once saved, the generated script can be imported into the live tables
using phpMyAdmin, or some other interface. It is hoped that the Admin
tool for importing SQL patches will also handle the script, but that
will require some testing as this script makes no allowances for the
possibility that the Zen-Cart store might have a database prefix set.

The same process of dual updating will be applied to the importing of
products tables and the products-categories placement table.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>