Update the products status in databases
Modified the status change detection to handle changes between states
in both directions. Filter the results to limit it to changes only.

Switch to using products_id rather than products_model for linking the
import and live data. Required adding a process to add the id values
to the import table.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>