Update product descriptions and names from imports
For products where a modification of the name or description has been
made by the admin, causing the `products_last_modified` column to be
non-null, no changes are made.

It is anticipated that any changes to the descriptions of the products
made by AzureGreen will be 'cosmetic' rather than substantive. If the
store admin has edited the information, it is presumed to be correct,
in substance and presentation, and updates from AzureGreen can be
ignored.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>