Add the new products from the import
Needed to have a category to mark products as new until they can be
further processed in the products to categories placement section.

Created a category for marking products which have issues to resolve
as well. Both categories will have to be added to the database in some
sort of *install* script. Setting them every time this script is run
seems a bit excessive. The same should be done for the AUTO_INCREMENT
values. Those are being kept here, until moved elsewhere, since they
__MUST__ be implemented from the beginning or errors an collisions
are inevitable. The missing categories will be trivial to add, without
consequences, later.

Changed INCREMENT_LIMIT to INCREMENT_BASE, as it is the *base* for the
ID values of items and categories added from the admin panel, not the
*limit* for such.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>