Add new product placements to the database
The earlier decision to not force the sorting category into the
database was in error. The ID for that category needs to be within the
range for non-AzureGreen categories, above `INCREMENT_BASE`, or it
will be removed from the table each time the process is run. Using an
ID in that range risks a collision with manually added categories if
it is not added to the database.

The creation of the categories themselves is left for some other place
to do.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>