Record non-leaf categories placements
Zen-Cart does not allow having categories and products mixed in the
same parent category. Either a category has sub categories or it has
products. AzureGreen is known to provide category placements which do
not follow that rule. There is no reliable method for automating the
correction of such placements, so the only option is to record it as
an error an allow the admin to alter the placement of the item.

At this point, I also do not see any method of "protecting" the manual
changes to placements from the next cycle of updates from AzureGreen's
data files. Added a condition to the `WHERE` clause for placements
that are added and dropped which _might_ protect such changes, but not
certain that it is enough.

As part of the above, the error table creation has been moved to the
setup section. Of the tables created in this process, they are the
only tables which are persistent.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>