Set the master categories for new products
The program calling the SQL script needs to create a supporting script
to set the `master_categories_id` values in the order preferred.

As the options are essentially infinite, it is easier to generate that
script under program control than to implement it here.

For each category, in the order desired, the script needs to do the
following, using `123` as the `categories_id` in the example:

    UPDATE `products`
    JOIN `products_to_categories`
        ON `products_to_categories`.`products_id`
            = `products`.`products_id`
    SET `master_categories_id`=123
    WHERE
        `master_categories_id`=@IMPORT_CATEGORY AND
        `categories_id`=123;

Any products not set by the supporting script are given one last shot
at a sane category. Any remaining products assigned the importing
category as their master category are reported in the errors table.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>