Compress the output from the SQL script
The output from the script is intended to be used to update the live
server with the imported data. Transmission time, and bandwidth, are
greatly saved by compressing the purely text file. It is best done
using an interface, such as phpMyAdmin, which can handle compressed
files. The admin interface sql patch tool cannot, and the user will
need to use the uncompressed version, which may be rather large.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>