Fix another oddity of AzureGreen
AzureGreen added two categories with the same name, 619 and 623,
"Viking Products" as subcategories of 617 "Viking." Of course, they
are only using 623 in the products placement, for now. To avoid issues
with the store-front, the 619 categories_id is being marked inactive,
until it's name is changed, or is removed by AzureGreen. If, or more
likely when, AzureGreen uses it as a placement category, the script
will have to be modded to switch any 619 placements to 623.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>