Drop temporary table usage
Removing the usage of temporary tables allows the script to be used on
a remote database that has a slow, often dropping connection. The user
will still be responsible for ensuring that the connection works, and
that all the processing is done.

It has been found that the persistence of the "temporary" tables can
be helpful in processing, or unscrambling, some of the mess made in
the files from AzureGreen.

This may, or may not, constitute an API change, but it will be called
on here. Version number will be bumped accordingly.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>