Add process for scraping missing images
AzureGreen only updates product images in the supplied zip files on a
monthly basis. This is far too long for a missing image to exist on
a Zen-Cart store. Missing images are usually available on the main
site for AzureGreen, as long as one is willing to use their browser to
retrieve them. That's too tedious for a regular work flow.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>