Handle spaces in supplied image file names
AzureGreen does not prevent spaces from being used in the file name of
images. This was causing the process to keep trying to find the file
and trying to download it at every check for missing images.

Stripped the space from the file name in the database, on the file
extracted from the zip file, and on scrapped images.

At the same time, forced all Jpeg files to be .jpg, even when the data
shows the name as a .jpeg file.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>