Allow for image file types other than Jpeg
With the unpredictable results so far from AzureGreen, it seems to be
a good idea to be ready for when they supply an image with some other
extension. Even invalid file types. The name, including extension,
will be used, so far as practical. Similar to the .jpeg to .jpg issue,
TIFF files will be forced to .tif file extension. Other extensions can
be left as-is, hoping they are correct. The pattern established for
folding the .jpeg and .tiff extensions into .jpg and .tif an be copied
to handle other folding, should the need be discovered.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>