Switch from wget to curl for file retrieval
The use of `curl` withing PHP is well established and the Zen-Cart
system has it built in alread. This will eventually need to be made
into a module for the store, and this is a small step to take now that
might make it easier then.

Change the image retrieval to HTTPS, as it is now supported by the
source.

Add a small amount of feedback reporting from the process to the user.
It was becomming obnoxious waiting for processing to finish. Especially
in the image archives when I had no idea where it is in the progress.

Change the granularity on the image archives to include the minutes.
After doing a second run in one day and having the file clobber the
old one, I decided to add that fix.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>