Restore file timestamping
The switch from `wget` to `curl` dropped the timestamping for the
downloaded files, instead giving them the current time. That is now
restored.

In addition, that timestamp is now used to name the archive directory
where new downloads are saved. This reflects the timing of changes as
they are made by AzureGreen better.

In concert with that, a new environment variable is looked for.
`RUN_DATE` can be set prior to the script running and will be
used as the date added for products found in that run of the script
if the script is reading historical data. New data will still have
the current date applied for new products.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>