Remove extra output to import script
Removed unneeded output into the importing script. The importing
script should not need to set the working categories and auto
increment values in the tables. That should have been done during the
setup, and if it was not there are likely to be problems later that
setting it now will not solve.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>