Add environment variable usage to shell script
The information used to connect to the database can now be set in
environment variables. If the variables are not set, or are blank, the
script will assign built-in defaults, if any. The script will also
verify that at least a database name and username are supplied. If the
script has a default password and one is not given in the environment
variables, the default will be used anyway. Ideally, the password will
be blank in both cases, avoiding the security risks associated with
plain text password storage.

README, and INSTALL files reflect the new status as well.

Signed-off-by: Ronald Lamoreaux <gitkey@chindraba.work>